backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
firebase-admin==6.6.0
gunicorn==23.0.0
numpy==2.2.1
python-dotenv==1.0.1
